pyproject.toml
# this is just for integration test purpose, it is not used for publishing to PyPI, we use setup.py
# for some manual test, you need to make a copy of this and rename it to ``pyproject.toml``
[tool.poetry]
name = "prompt_ide"
version = "0.1.1"
description = ""
authors = ["Sanhe Hu <husanhe@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<4.0" # this is just for integration test purpose, it doesn't mean that we only support 3.10+
sqlalchemy = ">=2.0.30,<3.0.0"
flask = ">=3.0.0,<4.0.0"
Werkzeug = ">=3.0.0,<4.0.0"
python-dateutil = ">2.8.2,<3.0.0"
MarkupSafe = ">=2.0.0,<3.0.0"
Flask-SQLAlchemy = ">=3.1.1,<4.0.0"
itsdangerous = ">=2.0.0,<3.0.0"
Flask-Restless-NG = ">=3.2.3,<3.3.0"


# --- addtitional dependencies for development
[tool.poetry.group.dev]
optional = true


[tool.poetry.group.dev.dependencies]
pathlib_mate = "*"
twine = "*"
wheel = "*"
build = "*"


# --- addtitional dependencies for building documentations
[tool.poetry.group.doc]
optional = true


[tool.poetry.group.doc.dependencies]
Sphinx = "5.3.0"
sphinx-jinja = "2.0.2"
sphinx-copybutton = "0.5.1"
sphinx-design = "0.5.0"
furo = "2023.03.27"
nbsphinx = "0.8.12"
rstobj = "1.2.1"
pygments = "2.15.1"
ipython = "8.10.0"
docfly = "2.0.1"


# --- addtitional dependencies for testing
[tool.poetry.group.test]
optional = true


[tool.poetry.group.test.dependencies]
pytest-cov = "*"
pytest = "*"
twine = "*"
wheel = "*"
build = "*"


# --- addtitional dependencies for automation shell scripts
[tool.poetry.group.auto]
optional = true

[tool.poetry.group.auto.dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
